Задание 7 - модуль 2
