+++ js/main.js
@@ -1,7 +1,2 @@
 import {picturesContainer} from './create-thumbnails.js';
-import {arrayPhotos} from '../js/create-array-photos.js';
-import {a} from '../js/create-big-picture.js';
-
-
-console.log(arrayPhotos);
